# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API Management",
    "description": "A comprehensive API for managing and interacting with a collection of public APIs. Allows users to view, add, update, and delete APIs, as well as retrieve detailed information on each API's functionality and usage."
  },
  "host": "publicapis.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/auth/google": {
      "tags": ["Authentication Routes"],
      "get": {
        "description": "Authenticate using Google",
        "responses": {
          "302": {
            "description": "Redirects to Google authentication"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/auth/google/callback": {
      "tags": ["Authentication Routes"],
      "get": {
        "description": "Google authentication callback",
        "responses": {
          "302": {
            "description": "Redirects to profile after successful authentication"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/profile": {
      "tags": ["User Routes"],
      "get": {
        "description": "Get user profile (authenticated only)",
        "responses": {
          "200": {
            "description": "User profile details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "User not authenticated"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/login": {
      "tags": ["Authentication Routes"],
      "get": {
        "description": "Login route for unauthenticated users",
        "responses": {
          "200": {
            "description": "Message to log in using Google"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/logout": {
      "tags": ["Authentication Routes"],
      "get": {
        "description": "Logout route for authenticated users",
        "responses": {
          "200": {
            "description": "Successfully logged out"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/apis": {
      "tags": ["API Routes"],
      "get": {
        "description": "Get all APIs",
        "responses": {
          "200": {
            "description": "Successfully fetched APIs"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "Add a new API",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                },
                "base_url": {
                  "type": "string"
                },
                "auth_required": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully created new API"
          },
          "400": {
            "description": "Invalid input or missing data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/apis/{id}": {
      "tags": ["API Routes"],
      "get": {
        "description": "Get a single API by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the API to fetch",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully fetched API"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "description": "Update an API by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the API to update",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                },
                "base_url": {
                  "type": "string"
                },
                "auth_required": {
                  "type": "boolean"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully updated API"
          },
          "400": {
            "description": "Invalid input or missing data"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Failed to update API"
          }
        }
      },
      "delete": {
        "description": "Delete an API by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the API to delete",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted API"
          },
          "404": {
            "description": "API not found"
          },
          "500": {
            "description": "Failed to delete API"
          }
        }
      }
    },
    "/users": {
      "tags": ["User Routes"],
      "get": {
        "description": "Get all users",
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "description": "Add a new user",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "googleId": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "googleLastName": {
                  "type": "string"
                },
                "googleName": {
                  "type": "string"
                }
              },
              "required": [
                "googleId",
                "email"
              ]
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfully created new user"
          },
          "400": {
            "description": "Invalid input or missing data"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/users/{id}": {
      "tags": ["User Routes"],
      "get": {
        "description": "Get user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "description": "Update a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "googleId": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "googleLastName": {
                  "type": "string"
                },
                "googleName": {
                  "type": "string"
                }
              },
              "required": [
                "googleId",
                "email"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User updated successfully"
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "description": "Delete a user by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The user ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User successfully deleted"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "googleId": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "googleName": {
            "type": "string"
          },
          "googleLastName": {
            "type": "string"
          }
        }
      }
    }
  }
}